+++ src/inventory.h
@@ -87,14 +87,14 @@ class Inventory final
         /**
          * Adds a new item in a free slot.
          */
-        void addItem(const int id, const int quantity, const int refine,
+        void addItem(const int id, const int quantity, const uint8_t refine,
                      const unsigned char color, const bool equipment = false);
 
         /**
          * Sets the item at the given position.
          */
         void setItem(const int index, const int id, const int quantity,
-                     const int refine, const unsigned char color,
+                     const uint8_t refine, const unsigned char color,
                      const bool equipment = false);
 
         /**
